+++ conf/env_test.go
@@ -1,0 +1,8 @@
+package conf
+
+import "testing"
+
+func TestEnv(t *testing.T) {
+	// TODO: LoadEnv
+	// TODO: LoadRequiredEnv
+}