+++ CHANGELOG.md
@@ -2,6 +2,10 @@
   * Commit: 
   * Fail functions
   * Conf package
+  * conf.LoadEnv
+  * dict.Inspect
+  * Sys package 
+  * Root package
 ## v0.3.6 - Schema Combos 
   * Commit: 2026-04-01 22:44
   * Schema GetOrCreate