+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## v0.3.8 - Root Package
-  * Commit: 
+  * Commit: 2026-04-07 15:13
   * New Commands
   * ParamsMap
   * Get command params
   * Display help and search command
+  * MainLoop
 ## v0.3.7 - Application Packages 
   * Commit: 2026-04-07 13:50
   * Fail functions